Fixed a few TODO's - Changed InternalWindowEvent queue to unique_ptr - Delete window in WindowGroup if group is stopping - Renamed Window shouldTerminate function to destroyAndFlagTerminated - Added clearing events after WindowGroup loop ended - Changed WindowPlugins to use unique_ptr in Config and Window - Finally fixed calculating Window delta time - Fixed setting vsync option - Moved GLFW_PLATFORM_X11 platform hint out of GLFWM and into example

+++ include/glfwm/Application.hpp
@@ -15,7 +15,7 @@ namespace nyx {
         explicit Application(WindowHandle &handle) : windowHandle(handle) {}
         virtual void configure(Config &config) = 0;
         virtual void create() = 0;
-        virtual void render(float dt, long frameTime) = 0;
+        virtual void render(double dt, long frameTime) = 0;
         virtual void resize(int width, int height) = 0;
         virtual void dispose() = 0;
         virtual ~Application() = default;
